+++ packages/react-native-sdk/ios/RTCViewPipManager.mm
@@ -12,5 +12,6 @@
 
 RCT_EXPORT_VIEW_PROPERTY(streamURL, NSString)
 RCT_EXTERN_METHOD(onCallClosed:(nonnull NSNumber*) reactTag)
+RCT_EXTERN_METHOD(setPreferredContentSize:(nonnull NSNumber *)reactTag width:(CGFloat)w height:(CGFloat)h);
 
 @end